nixos/anki-sync-server: fix default SYNC_HOST
SYNC_HOST isn't a host name, it must be an address. 'localhost' fails with:
```
Sync server failed: reading SYNC_* env vars

Caused by this error:
  1: invalid IP address syntax
```

This is true of both anki --syncserver and anki-sync-server.

Rename the nixos services config to 'address' to make this clear and
improve doc for it.