bpftrace: fix tcp*.bt scripts
These scripts tried to include glibc headers, which didn't work on nixos
because bpftrace tried to call clang-major with some options to get the
paths, but we never told it where to find it.
Give bpftrace the path directly instead of giving it a path to clang.

Note the default clang output also had some clang-internal headers, but
these seem to be included anyway through libclang API without having to
give them here, so we only pass the glibc ones.

While here:
- also backport another bpftrace patch that made the tcp*.bt tools to
actually not require these includes
- and add tests that we can include these.

Co-authored-by: Arian van Putten <arian.vanputten@gmail.com>